% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildpops.R
\name{buildpops}
\alias{buildpops}
\title{buildpops}
\usage{
buildpops(gpfile, pop.file)
}
\arguments{
\item{gpfile}{Name of genepop file. REQUIRED.}

\item{pop.file}{Name of table with population designation, 1 pop per row. REQUIRED.}
}
\description{
Sort, filter, and rearrange genepop populations by individual names
}
\examples{
buildpops (gpfile="Populations.gen", pop.file="5_pops.txt")
}
\keyword{buildpops}
