% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixformat.R
\name{fixformat}
\alias{fixformat}
\title{fixformat}
\usage{
fixformat(gpfile)
}
\arguments{
\item{gpfile}{Genepop file. REQUIRED.}
}
\description{
Fixes common whitespace and invalid character errors in genepop files
}
\examples{
fixformat(gpfile="Populations.gen")
}
\keyword{fixformat}
