% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genosim.R
\name{genosim}
\alias{genosim}
\title{genosim}
\usage{
genosim(infile, iter = 100, hwefile = NULL)
}
\arguments{
\item{infile}{Input allele frequency table (from GPoppin::popgen). REQUIRED.}

\item{iter}{Number of individuals to simulate per population. Default = 100.}

\item{hwefile}{Input Hardy-Weinberg disequilbrium table (from GPoppin::popgen) Default = NULL.}
}
\description{
Simulate Genepop genotypes using allele frequency tables and HW disequilibrium
}
\examples{
genosim(infile="Populations_freqtable.txt", iter = 200, hwefile = "Populations_loc_dis.txt")
}
\keyword{genosim}
