% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchloci.R
\name{matchloci}
\alias{matchloci}
\title{matchloci}
\usage{
matchloci(gpfile, loci.file)
}
\arguments{
\item{gpfile}{Genepop filename to be reorganized. REQUIRED.}

\item{loci.file}{Filename of list of loci in desired order. REQUIRED.}
}
\description{
Reorganizes loci in Genepop file to match the order of another list of loci.
}
\examples{
matchloci(gpfile="Population.gen", loci.file="Loci_by_significance.txt")
}
\keyword{matchloci}
