% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{popgen}
\alias{popgen}
\title{popgen}
\usage{
popgen(gpfile, maf = TRUE, geno.s = TRUE, fail.pops = TRUE,
  pop.thres = 1, loc.stats = TRUE, maf.fz = 0.05, fail.loci = TRUE,
  pop.table = TRUE, hwe.calc = FALSE, iter = 100, p.pop = 0.33,
  p.thres = 0.05, calc.fst = FALSE, pca = TRUE, fz.table = TRUE)
}
\arguments{
\item{gpfile}{Genepop filename. REQUIRED.}

\item{maf}{Create minor allele frequency table for each locus. Default=TRUE.}

\item{geno.s}{Determine genotype success of each locus. Default=TRUE.}

\item{fail.pops}{Flag loci that aren't sequenced in at least N population. Default=TRUE.}

\item{pop.thres}{Number of populations a locus must be present in. If fail.pops = TRUE. Default = 1.}

\item{loc.stats}{Calculate stats such as He, Ho, He-Ho, FIS for each locus. Default = TRUE.}

\item{maf.fz}{Flag loci that have a minor allele frequency less than N. Default = 0.05.}

\item{pop.table}{Create a summary table of population level analyses (Ho,He, Ho-He, FIS, GenoS,N). Default=TRUE.}

\item{hwe.calc}{Calculate Hardy-Weinberg equilibrium for each locus. Default = FALSE.}

\item{iter}{Number of iterations to perform in HWE analysis. If hwe.calc = TRUE. Default = 100.}

\item{p.pop}{Flag loci that are out of HWE in N proportion of populations. If hwe.calc = TRUE. Default = 0.33.}

\item{p.thres}{p-value threshold for HWE tests. If hwe.calc = TRUE. Default = 0.05.}

\item{calc.fst}{Calculate FST matrix (by population). Default = FALSE.}

\item{pca}{Perform an individual-level principal component analysis. Default = TRUE.}

\item{fz.table}{Create an allele frequency table for each locus. Default = TRUE.}
}
\description{
Perform multiple population genetic analyses on a Genepop file and flag loci that don't meet criteria
}
\examples{
popgen(gpfile="Populations.gen", maf= TRUE, geno.s= TRUE, fail.pops= TRUE, pop.thres=1,
loc.stats= TRUE, maf.fz=0.05, pop.table= TRUE, hwe.calc=TRUE,
iter=100, p.pop=0.33, p.thres=0.05, calc.fst= TRUE, pca=TRUE, fz.table= TRUE)
}
\keyword{popgen}
