\name{print.boostmtree}
\alias{print.boostmtree}
\title{Print Summary Output}
\description{
    Print summary output from the boosting analysis.
}
\usage{\method{print}{boostmtree}(x, ...)}
\arguments{
    \item{x}{An object of class \code{(boostmtree, grow)} or \code{(boostmtree,
	predict)}.}
    \item{...}{Further arguments passed to or from other methods.}
}
\author{
    Hemant Ishwaran and Udaya B. Kogalur
}
\references{
  Ishwaran et al. (2014).  Boosting multivariate trees for
  longitudinal data.
}
\keyword{print}
