% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prog2gp.R
\name{prog2gp}
\alias{prog2gp}
\title{prog2gp}
\usage{
prog2gp(infile, rem.indels = FALSE, rem.sex = FALSE, fix.loc = TRUE,
  rem.thres = 0, split.pop = FALSE, rem.dup = TRUE, first.col = FALSE,
  examp = FALSE)
}
\arguments{
\item{infile}{Input progeny export table. csv or txt. REQUIRED.}

\item{rem.indels}{Remove indels from dataset. Default = FALSE.}

\item{rem.sex}{Remove sex markers from dataset. Default = FALSE.}

\item{fix.loc}{Fix problematic locus names that contain invalid characters for R packages. Default = TRUE.}

\item{rem.thres}{Remove loci that are not genotyped in N percent of individuals. Default = 0.}

\item{split.pop}{Attempt to split Genepop file into pops based on individual names. Default = FALSE.}

\item{rem.dup}{Remove duplicated individuals by name. Default = TRUE.}

\item{first.col}{Genotypes start at the second column, after individual names. Default = FALSE.}

\item{examp}{Use example file. Default = FALSE.}
}
\description{
Convert a progeny export csv, 1 allele per column format, to Genepop format.
}
\examples{
prog2gp(infile = "progeny_export.csv", rem.indels=FALSE, rem.sex=FALSE,
fix.loc=TRUE, rem.thres=0, split.pop=TRUE, rem.dup=TRUE, first.col = FALSE, examp=FALSE)
}
\keyword{prog2gp}
