% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortloci.R
\name{sortloci}
\alias{sortloci}
\title{sortloci}
\usage{
sortloci(gpfile, locus.file, option)
}
\arguments{
\item{gpfile}{Name of Genepop file that will be filtered. REQUIRED.}

\item{locus.file}{Name of file that is a list of locus names. REQUIRED.}

\item{option}{Indicate if you want to keep or remove loci in the locus.file. Either "keep" or "remove". REQUIRED.}
}
\description{
Keep or remove a list of loci from a Genepop file
}
\examples{
sortloci(gpfile = "Population.gen", locus.file= "significant_snps.txt", option = "keep")
}
\keyword{sortloci}
