\name{print.boostmtree}
\alias{print.boostmtree}
\title{Print Summary Output of a RF-SRC Analysis}
\description{
    Print summary output from a RF-SRC analysis. This is the default
    print method for the package.
}
\usage{\method{print}{boostmtree}(x, ...)}
\arguments{
    \item{x}{An object of class \code{(rfsrc, grow)} or \code{(rfsrc,
	predict)}.}
    \item{...}{Further arguments passed to or from other methods.}
}
\author{
    Hemant Ishwaran and Udaya B. Kogalur
}
\references{
    Ishwaran H. and Kogalur U.B. (2007).  Random survival forests for R,
    \emph{Rnews}, 7/2:25-31.
}
\seealso{
  \command{\link{predict.rfsrc}}
}
\examples{
}
\keyword{print}
